Shader: Refine preprocessing on GLES platform
Don't globally set float precision to mediump, only fragment
shaders need that and defining it for vertex shaders causes
tilemap cracks.

Also manually define low precision for variables that hold
color / alpha values.

+++ shader/bitmapBlit.frag
@@ -6,7 +6,7 @@ uniform sampler2D destination;
 
 uniform vec4 subRect;
 
-uniform float opacity;
+uniform lowp float opacity;
 
 varying vec2 v_texCoord;
